Support hijacking of TLS listeners
