+++ server.go
@@ -47,7 +47,6 @@ import (
 	"net/http"
 	"sync"
 	"sync/atomic"
-	"time"
 )
 
 // interface describing a waitgroup, so unit
@@ -147,9 +146,8 @@ func (s *GracefulServer) ListenAndServeTLSWithConfig(config *tls.Config) error {
 			return err
 		}
 
-		tlsListener := tls.NewListener(TCPKeepAliveListener{ln.(*net.TCPListener)}, config)
+		tlsListener := NewTLSListener(TCPKeepAliveListener{ln.(*net.TCPListener)}, config)
 		s.listener = NewListener(tlsListener)
-
 	}
 	return s.Serve(s.listener)
 }
@@ -161,7 +159,7 @@ func (gs *GracefulServer) HijackListener(s *http.Server, config *tls.Config) (*G
 	}
 
 	if config != nil {
-		listener = tls.NewListener(TCPKeepAliveListener{listener.(*net.TCPListener)}, config)
+		listener = NewTLSListener(TCPKeepAliveListener{listener.(*net.TCPListener)}, config)
 	}
 
 	other := NewWithServer(s)
@@ -299,24 +297,4 @@ func Close() {
 	}
 	servers = nil
 	m.Unlock()
-}
-
-// TCPKeepAliveListener sets TCP keep-alive timeouts on accepted
-// connections. It's used by ListenAndServe and ListenAndServeTLS so
-// dead TCP connections (e.g. closing laptop mid-download) eventually
-// go away.
-//
-// direct lift from net/http/server.go
-type TCPKeepAliveListener struct {
-	*net.TCPListener
-}
-
-func (ln TCPKeepAliveListener) Accept() (c net.Conn, err error) {
-	tc, err := ln.AcceptTCP()
-	if err != nil {
-		return
-	}
-	tc.SetKeepAlive(true)
-	tc.SetKeepAlivePeriod(3 * time.Minute)
-	return tc, nil
 }
